Add files via upload
10-bit Single-Cycle CPU
Lab objectives:
The main objective of this lab is to give the student a deeper understanding of singlecycle-
architecture implementation and design tradeoffs. In this lab, students will learn
how to design a control unit for single-cycle architecture, and connect the ALU, register
file and fetch unit they implemented in previous lab to each other. This laboratory will
also allow students to understand how their design choices affect the execution time
and hardware complexity. The project will additionally enhance their ability of critical
thinking and working in teams.
Lab details:
In this lab, you will design a single-cycle microarchitecture. You will use your fetch unit,
ALU, and registers you designed in lab2. You will implement a combinational logic
control unit and integrate it to your lab2. You will also need to connect everything
together to make a single-cycle 10-bit CPU. In this lab, you will not implement an
instruction and data memory yet. You will use Xilinx Vivado tools for your design.
You will need to design and implement the following:
1- Design a combinational logic control unit module. Try to isolate your control in
one or few modules if you can and not spread it inside other modules, such as
ALU or register file.
2- Connect all of your modules from lab2 (ALU, Fetch Unit and Register file and any
special register), and the control unit in one main module (CPU10bits).
3- Find the longest-path cycle time--assuming each instruction executes in one
cycle for you CPU10bits top module. Use this cycle time as your clock period
time in your simulation test files.
4- The input to your main module, CPU10bits, should be one clock signal, one reset
signal and 10-bit instruction input. The output should be one done bit. The done
bit is set to 1 whenever a halt instruction executes, and to 0 otherwise.
5- Synthesize, and show your schematic and timing diagrams for each new module
you implement in this lab and each module you modified from lab2 separately.
6- Test all of your instructions that exist in your ISA and show a timing diagram for
different test cases for instructions. Although you are not implementing data
memory in this part of the lab, you need to test your load and store instructions,
and show intermediate results such as memory addresses.
7- In addition to 4, Implement the third program from your lab 1 (f = (x*y) - 4 ), by
encoding all your instructions and inputting them to the CPU10bit. Show the
results and internal state of the design in the timing diagram. Make sure your
program takes in consideration the different possible signs of x and y. The easiest
way is to remember the sign and convert negative number to positive one then
multiply. If the sign of both was different then the result will be converted to
negative.