Add files via upload
10-bit Pipelined CPU
Lab objectives:
The main objective of this lab is to give students a deeper understanding of a multi-cycle
architecture implementation and design tradeoffs. In this lab, students will learn how to
design a pipeline for their own 10-bit CPU. This laboratory will also allow students to
understand how their design choices affect clock period time, execution time and
hardware complexity. The project will additionally enhance their ability of critical
thinking and working in teams.
Lab details:
In this lab, you will design a 3 stage-pipelined microarchitecture. The three stages are:
Fetch-decode, execute-memory, and write back stages. You will first design pipeline
registers that store the relevant information for each instruction moving from one stage
to the other in the pipeline including relevant control signals. Then you will need to split
your single cycle microarchitecture from Lab3B into three stages; you can create three
different modules to make debugging easier. Finally, you will integrate your stages (or
modules) and the two pipeline registers. You will need to run the three programs you
wrote for lab1 on your 10-bit pipelined CPU, by storing them in the instruction memory
one at a time. You will use Xilinx Vivado tools for your design.
Implementation details
You will need to design and implement the following:
1- Fetch-decode stage should include all your: ROM, fetch unit, control unit and
register file modules. Branch execution logic should be included in this stage as
well.
2- Execute-memory stage should include your, ALU and data memory and any
special purpose flags or registers that your design set when an instruction is
executed.
3- Write back stage should include the logic to write back the results to the register
file (note register file implementation is in the decode stage so you do not read
to instantiate). Your register file should be read in the second half of the cycle
and written to in the first half of the cycle to avoid any structural hazard.
4- Design two pipeline registers to fit all the information you need to pass for any
instruction type to execute correctly and perform what is required for
instructions on each pipeline stage. The first pipeline register is fetchdecode/
execute-memory, and the second is execute-memory/writeback
register. Note that you need to always store and pass to each stage what is
required to execute any instruction type for that particular stage. You may need
to store more information than is needed for the current instruction being
executed. In other words, assume the worst case. Note that the two pipeline
registers do not have to have the same size as each stage may require different
information.
5- Your design should allow register file to be read from the fetch-decode stage (in
the second half of the cycle) and written in the writeback stage in the first half of
the clock cycle.
6- Integrate your modules to form a 10-bit pipelined CPU (modular integration and
testing is recommended).
7- Find the longest-path cycle time for each stage. Then find the longest time for all
stages for your CPU10bits top module to determine the clock period time. Use
this cycle time as your clock period time in your simulation test files.
8- The input to your main module, CPU10bits, will be one clock signal and one reset
signal. The output to your CPU module should be one done that is set to ‘1’
when a halt instruction is executed and set to ‘0’ otherwise. In addition to the
done bit, add a 10-bit output that show the value of the result of the instruction after it
finished writing back its results in the write back stage. If the instruction does not
produce a result (e.g a branch instruction), then this output should be set to all X’s.
9- Synthesize, and show your schematic and timing diagrams for your 10-bit
pipelined CPU and for each new module you implement in this lab and each
module you modified from lab3b separately.
10- Test all your three programs with different data inputs.
11- Your design needs to make sure that instructions that depends on each other
execute correctly. See the following.
For your programs to work correctly you will need to check for data dependency. You
will need to design:
Forwarding module that forwards (bypass) the result from the execute-memory
stage of one instruction to the next in case of data dependency. The forwarding
unit needs to compare the destination of one instruction (stored in the executememory/
write back pipeline register) to the sources of the next instruction
(stored in the fetch-decode/execute-memory pipeline register) and forward the
result to the designated operand in case of a match. You should include an
enable signal to enable forwarding or disable it. After implementing the
forwarding module repeat steps 5 to 9 above. Your CPU design should include
one more input signal, in addition to what is indicated in 8 above, called
“forward” to enable forwarding by the user when needed.